feat: Update NotionTableView component to handle phone number and email updates
This commit updates the NotionTableView component to handle phone number and email updates in the Notion API. It adds the functionality to update the phone number and email fields for a given page ID. If the update is successful, the status code 202 is returned. If there is an error, an error message is logged to the console. These changes improve the functionality of the component and ensure that the correct data is saved in the Notion API.