Update condition and label for Sidebar Link
Added logic to display 'Marketplace' link based on the website's URL in the Sidebar component. The 'MarketPlace' link will now be visible only when the app isn't running on specified production or development domains, improving clarity and navigation for external users. The label for the link has also been updated from 'Ativos Disponíveis' to 'MarketPlace' for better understanding and consistency with other parts of the application.