feat: Add MarketplaceItem component and page
This commit adds a new component called MarketplaceItem, which is used to display a specific item in the marketplace. It also includes a new page called MarketplaceItemPage, which uses the MarketplaceItem component to render the item details. This commit is a feature addition to enhance the functionality of the application.